chore: consolidate deduplicateResults
