+++ src/__tests__/linter.jest.test.ts
@@ -65,12 +65,12 @@ describe('Linter', () => {
     await spectral.loadRuleset(customDirectoryFunctionsRuleset);
     expect(await spectral.run({})).toEqual([
       expect.objectContaining({
-        code: 'has-info-property',
-        message: 'info property is missing',
-      }),
-      expect.objectContaining({
         code: 'has-field-property',
         message: 'Object does not have field property',
+      }),
+      expect.objectContaining({
+        code: 'has-info-property',
+        message: 'info property is missing',
       }),
     ]);
   });
