+++ src/__tests__/linter.test.ts
@@ -162,15 +162,15 @@ describe('linter', () => {
 
     expect(result).toEqual([
       expect.objectContaining({
-        code: 'invalid-ref',
-      }),
-      expect.objectContaining({
-        code: 'invalid-ref',
-      }),
-      expect.objectContaining({
         code: 'oas3-schema',
         message: "/paths//pets/get/responses/200 should have required property '$ref'",
         path: ['paths', '/pets', 'get', 'responses', '200'],
+      }),
+      expect.objectContaining({
+        code: 'invalid-ref',
+      }),
+      expect.objectContaining({
+        code: 'invalid-ref',
       }),
       expect.objectContaining({
         code: 'oas3-unused-components-schema',
@@ -605,7 +605,7 @@ responses:: !!foo
     );
   });
 
-  test('should remove all redundant ajv errors', async () => {
+  test.only('should remove all redundant ajv errors', async () => {
     spectral.registerFormat('oas2', isOpenApiv2);
     spectral.registerFormat('oas3', isOpenApiv3);
     await spectral.loadRuleset('spectral:oas');
@@ -614,26 +614,21 @@ responses:: !!foo
 
     expect(result).toEqual([
       expect.objectContaining({
-        code: 'invalid-ref',
-      }),
-      expect.objectContaining({
-        code: 'invalid-ref',
+        code: 'openapi-tags',
       }),
       expect.objectContaining({
         code: 'operation-tag-defined',
       }),
       expect.objectContaining({
-        code: 'openapi-tags',
-      }),
-      expect.objectContaining({
-        code: 'oas3-valid-schema-example',
-        message: '"foo.example" property type should be number',
-        path: ['components', 'schemas', 'foo', 'example'],
-      }),
-      expect.objectContaining({
         code: 'oas3-schema',
         message: "/paths//pets/get/responses/200 should have required property '$ref'",
         path: ['paths', '/pets', 'get', 'responses', '200'],
+      }),
+      expect.objectContaining({
+        code: 'invalid-ref',
+      }),
+      expect.objectContaining({
+        code: 'invalid-ref',
       }),
       expect.objectContaining({
         code: 'oas3-unused-components-schema',
@@ -644,6 +639,11 @@ responses:: !!foo
         code: 'oas3-unused-components-schema',
         message: 'Potentially unused components schema has been detected.',
         path: ['components', 'schemas', 'foo'],
+      }),
+      expect.objectContaining({
+        code: 'oas3-valid-schema-example',
+        message: '"foo.example" property type should be number',
+        path: ['components', 'schemas', 'foo', 'example'],
       }),
     ]);
   });
