+++ src/cli/services/linter/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './deduplicateResults';
 export * from './getRuleset';
 export * from './listFiles';
 export * from './skipRules';
