+++ src/spectral.ts
@@ -1,12 +1,10 @@
-const md5 = require('blueimp-md5');
-
 import { getLocationForJsonPath as getLocationForJsonPathJson, JsonParserResult, safeStringify } from '@stoplight/json';
 import { Resolver } from '@stoplight/json-ref-resolver';
 import { ICache, IUriParser } from '@stoplight/json-ref-resolver/types';
 import { extname, normalize } from '@stoplight/path';
 import { DiagnosticSeverity, Dictionary, IDiagnostic, Optional } from '@stoplight/types';
 import { getLocationForJsonPath as getLocationForJsonPathYaml, YamlParserResult } from '@stoplight/yaml';
-import { memoize, merge, uniqBy } from 'lodash';
+import { memoize, merge } from 'lodash';
 
 import { STATIC_ASSETS } from './assets';
 import { formatParserDiagnostics, formatResolverErrors } from './error-messages';
@@ -33,7 +31,7 @@ import {
   RunRuleCollection,
 } from './types';
 import { IRuleset } from './types/ruleset';
-import { ComputeFingerprintFunc, defaultComputeResultFingerprint, empty } from './utils';
+import { ComputeFingerprintFunc, defaultComputeResultFingerprint, empty, prepareResults } from './utils';
 
 memoize.Cache = WeakMap;
 
@@ -119,19 +117,9 @@ export class Spectral {
 
     validationResults = validationResults.concat(runRules(resolved, this.rules, this.functions));
 
-    // add fingerprint func to each result, to uniquely identify and group them
-    const computeFingerprint = this._computeFingerprint;
-    for (const r of validationResults) {
-      Object.defineProperty(r, 'fingerprint', {
-        get() {
-          return computeFingerprint(r, md5);
-        },
-      });
-    }
-
     return {
       resolved: resolved.resolved,
-      results: uniqBy(validationResults, 'fingerprint'),
+      results: prepareResults(validationResults, this._computeFingerprint),
     };
   }
