+++ src/utils/index.ts
@@ -2,4 +2,4 @@ export * from './empty';
 export * from './hasIntersectingElement';
 export * from './isObject';
 export * from './refs';
-export * from './fingerprint';
+export * from './prepareResults';
